// frontend/src/models/team/ITeamStatistic.ts
export interface ITeamStatistic {
    rank: number
    team_id: number
    team_name: string
    wins: number
    losses: number
    draws: number
    total_points: number
    win_percentage: number
    season_id: number
    season_name: string
    week_id: number
    week_start_date: string
    week_end_date: string
}